Refactor vertex buffer filters
Instead of passing the filter to decodeVertexBuffer as a callback, call
filter externally.

While in theory, this may be slower because the data is unlikely to be
in L1/L2 cache, in practice this only matters when decoding runs at
memory speed or faster, and it's not.

Removing the callback support makes the JS-C boundary much simpler and
more cleanly separates GLTF support code from the raw encoders - outside
of GLTF the recommended approach is to use octahedral encoding for
normals anyway.

This change also adds decoder.decodeGltfBuffer to simplify the filter
management so that both three.js and Babylon implementations can just
call one function.