vertexcodec: Use fast float->int conversion
Instead of using trunc_sat instruction that expands into 7 SSE
instructions, we can use the old trick of adding a large value to a
floating point number to force the mantissa bits to be equal to the
integer value of the original number.

This automatically rounds to nearest integer which is exactly what we
need so we can also save an addition. As a result, decoding normal
stream goes from 845 MB/s to 920 MB/s.